DOMA-1356 move ResidentActions.tsx outside menu component
