--- apps/condo/domains/common/components/containers/BaseLayout/TopMenuOnlyLayout.tsx
@@ -1,19 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-nocheck
-import React from 'react'
-
-import BaseLayout, { PageContent, PageWrapper } from './BaseLayout'
-
-function TopMenuOnlyLayout ({ children, ...props }) {
-    return <BaseLayout
-        {...props}
-        logoLocation="topMenu"
-        className="top-menu-only-layout"
-    >
-        <PageWrapper>
-            <PageContent>{children}</PageContent>
-        </PageWrapper>
-    </BaseLayout>
-}
-
-export default TopMenuOnlyLayout
