+++ apps/condo/domains/common/components/containers/BaseLayout/index.ts
@@ -1,9 +1,7 @@
 import BaseLayout, { PageContent, PageHeader, PageWrapper, useLayoutContext } from './BaseLayout'
-import TopMenuOnlyLayout from './TopMenuOnlyLayout'
 
 export default BaseLayout
 export {
-    TopMenuOnlyLayout,
     useLayoutContext,
     PageWrapper,
     PageHeader,
